# Improves performance EDN parse by ~80% by using std::str::Chars