--- a/Makefile
+++ b/Makefile
@@ -1,12 +1,13 @@
-int:
+integration:
 	cargo test --test lib --no-fail-fast --features "json"
 
 unit:
 	cargo test --locked  --no-fail-fast --lib
 
-ex:
+.PHONY: examples
+examples:
 	cargo test --examples
 	cargo test --example json_to_edn --features "json"
 	cargo run --example async --features "async"
 
-test: unit int ex
+test: unit integration examples