```
Benchmarking parse: Collecting 100 samples in estimated 5.0323 s (439k iteration                                                                              
parse                   time:   [13.570 us 13.606 us 13.641 us]
Performance has improved: 80.560%
```

We still can improve to 5us

PLEASE REVIEW BUT DONT MERGE IT