@otaviopace 2020-09-06
Great PR :clap: :clap: :clap: :clap: :clap:
Congrats on making the parsing more reliable!! :blush: 
I just added a couple notes on some things I think can be influencing performance, can be simplified and or could cause problems :wink: